+++ KasiCornerKota_Application/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantCommandHandler.cs
@@ -1,5 +1,7 @@
-﻿using KasiCornerKota_Domain.Entities;
+﻿using KasiCornerKota_Domain.Constants;
+using KasiCornerKota_Domain.Entities;
 using KasiCornerKota_Domain.Exceptions;
+using KasiCornerKota_Domain.Interfaces;
 using KasiCornerKota_Domain.Repositories;
 using MediatR;
 using Microsoft.Extensions.Logging;
@@ -8,7 +10,7 @@ using Microsoft.Extensions.Logging;
 namespace KasiCornerKota_Application.Restaurants.Commands.DeleteRestaurant
 {
     public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandler> logger,
-        IRestaurantsRepository restaurantsRepository) : IRequestHandler<DeleteRestaurantCommand>
+        IRestaurantsRepository restaurantsRepository, IRestaurantAuthorizationService restaurantAuthorization) : IRequestHandler<DeleteRestaurantCommand>
     {
         public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
         {
@@ -17,6 +19,12 @@ namespace KasiCornerKota_Application.Restaurants.Commands.DeleteRestaurant
 
             if (restaurant is null)
                 throw new NotFoundException(nameof(Restaurant), request.Id.ToString());
+
+
+            if(!restaurantAuthorization.Authorize(restaurant, ResourceOperation.Delete))
+            {
+                throw new ForbidException("You are not authorized to delete this restaurant.");
+            }
 
             await restaurantsRepository.Delete(restaurant);
